[9.3.0] Split off `TypesTest` from `TypeResolverTest`
This refactors `TypeResolverTest` to be more focused on `TypeResolver`'s behavior. A test of `Types.callable()` is moved to a new suite (currently in `syntax/`). The remaining test cases are renamed to match the method naming scheme in the tested class ("resolve..." -> "eval...").

Work toward #27728.

PiperOrigin-RevId: 842857557

Commit https://github.com/bazelbuild/bazel/commit/9fbbf3967eb95033b9f06549ae42e3eb9852ac47
