+++ src/main/java/net/starlark/java/types/StarlarkType.java
@@ -20,15 +20,19 @@ import java.util.List;
 /**
  * Base class for all Starlark types.
  *
- * <p>Tracking issue: https://github.com/bazelbuild/bazel/issues/22935
+ * <p>Starlark typing is an experimental feature under development. See the tracking issue:
+ * https://github.com/bazelbuild/bazel/issues/27370
  */
 public abstract class StarlarkType {
+
   /**
    * Returns the list of supertypes of this type.
    *
    * <p>Preferred order is from the most specific to the least specific supertype. But if that is
    * not possible, the order can be arbitrary.
    */
+  // TODO: #27370 - Add getSubtypes(), with the semantics that the actual subtype relation is the
+  // union of these two methods.
   public List<StarlarkType> getSupertypes() {
     return ImmutableList.of();
   }
