+++ src/test/java/net/starlark/java/syntax/TypeResolverTest.java
@@ -17,8 +17,6 @@ package net.starlark.java.syntax;
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.assertThrows;
 
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableSet;
 import net.starlark.java.types.StarlarkType;
 import net.starlark.java.types.Types;
 import org.junit.Test;
@@ -28,8 +26,16 @@ import org.junit.runners.JUnit4;
 /** Tests for Starlark type resolution. */
 @RunWith(JUnit4.class)
 public class TypeResolverTest {
+
+  private StarlarkType evalType(String type) throws Exception {
+    Expression typeExpr = Expression.parseTypeExpression(ParserInput.fromLines(type));
+    // TODO: #27728 - When type resolution can consider non-universal types, use a better mock
+    // module here that supports evalType().
+    return TypeResolver.evalTypeExpression(typeExpr, Resolver.moduleWithPredeclared());
+  }
+
   @Test
-  public void resolveType_onPrimitiveTypes() throws Exception {
+  public void evalType_primitives() throws Exception {
     assertThat(evalType("None")).isEqualTo(Types.NONE);
     assertThat(evalType("bool")).isEqualTo(Types.BOOL);
     assertThat(evalType("int")).isEqualTo(Types.INT);
@@ -38,12 +44,15 @@ public class TypeResolverTest {
   }
 
   @Test
-  public void resolveType_union() throws Exception {
+  public void evalType_union() throws Exception {
     assertThat(evalType("int|bool")).isEqualTo(Types.union(Types.INT, Types.BOOL));
   }
 
+  // TODO: #27370 - Rather than test applications of constructors for list and dict here, test the
+  // general machinery for calling a type constructor. The actual types should be tested separately.
+
   @Test
-  public void resolveType_list() throws Exception {
+  public void evalType_list() throws Exception {
     assertThat(evalType("list[int]")).isEqualTo(Types.list(Types.INT));
     assertThat(evalType("list[list[int]]")).isEqualTo(Types.list(Types.list(Types.INT)));
 
@@ -52,10 +61,12 @@ public class TypeResolverTest {
 
     exception = assertThrows(SyntaxError.Exception.class, () -> evalType("list[[int]]"));
     assertThat(exception).hasMessageThat().isEqualTo("unexpected expression '[int]'");
+
+    // TODO: #27370 - `list` should produce `list[Any]`.
   }
 
   @Test
-  public void resolveType_dict() throws Exception {
+  public void evalType_dict() throws Exception {
     assertThat(evalType("dict[int, str]")).isEqualTo(Types.dict(Types.INT, Types.STR));
     assertThat(evalType("dict[int, list[str]]"))
         .isEqualTo(Types.dict(Types.INT, Types.list(Types.STR)));
@@ -74,17 +85,18 @@ public class TypeResolverTest {
     assertThat(exception)
         .hasMessageThat()
         .isEqualTo("expected type arguments after the type constructor 'dict'");
+    // TODO: #27370 - `dict` should produce `dict[Any, Any]`.
   }
 
   @Test
-  public void resolveType_unknownIdentifier() {
+  public void evalType_unknownIdentifier() {
     SyntaxError.Exception e = assertThrows(SyntaxError.Exception.class, () -> evalType("Foo"));
 
     assertThat(e).hasMessageThat().isEqualTo("type 'Foo' is not defined");
   }
 
   @Test
-  public void resolveType_badTypeApplications() {
+  public void evalType_badTypeApplications() {
     SyntaxError.Exception e =
         assertThrows(SyntaxError.Exception.class, () -> evalType("int[bool]"));
     assertThat(e)
@@ -98,85 +110,5 @@ public class TypeResolverTest {
     assertThat(e)
         .hasMessageThat()
         .isEqualTo("expected type arguments after the type constructor 'list'");
-  }
-
-  // TODO: #27728 - This is a test of CallableType, not the TypeResolver. Split into its own file,
-  // either in this package or in types/.
-  @Test
-  public void callable_toSignatureString() {
-    // ordinary only
-    assertThat(
-            Types.callable(
-                    /* parameterNames= */ ImmutableList.of("a"),
-                    /* parameterTypes= */ ImmutableList.of(Types.INT),
-                    /* numPositionalOnlyParameters= */ 0,
-                    /* numPositionalParameters= */ 1,
-                    /* mandatoryParams= */ ImmutableSet.of("a"),
-                    /* varargsType= */ null,
-                    /* kwargsType= */ null,
-                    Types.BOOL)
-                .toSignatureString())
-        .isEqualTo("(a: int) -> bool");
-    // kwonly only
-    assertThat(
-            Types.callable(
-                    /* parameterNames= */ ImmutableList.of("a"),
-                    /* parameterTypes= */ ImmutableList.of(Types.INT),
-                    /* numPositionalOnlyParameters= */ 0,
-                    /* numPositionalParameters= */ 0,
-                    /* mandatoryParams= */ ImmutableSet.of("a"),
-                    /* varargsType= */ null,
-                    /* kwargsType= */ null,
-                    Types.BOOL)
-                .toSignatureString())
-        .isEqualTo("(*, a: int) -> bool");
-    // positional-only only
-    assertThat(
-            Types.callable(
-                    /* parameterNames= */ ImmutableList.of("x"),
-                    /* parameterTypes= */ ImmutableList.of(Types.INT),
-                    /* numPositionalOnlyParameters= */ 1,
-                    /* numPositionalParameters= */ 0,
-                    /* mandatoryParams= */ ImmutableSet.of(),
-                    /* varargsType= */ null,
-                    /* kwargsType= */ null,
-                    Types.BOOL)
-                .toSignatureString())
-        .isEqualTo("([int], /) -> bool");
-    // no params
-    assertThat(
-            Types.callable(
-                    /* parameterNames= */ ImmutableList.of(),
-                    /* parameterTypes= */ ImmutableList.of(),
-                    /* numPositionalOnlyParameters= */ 0,
-                    /* numPositionalParameters= */ 0,
-                    /* mandatoryParams= */ ImmutableSet.of(),
-                    /* varargsType= */ null,
-                    /* kwargsType= */ null,
-                    Types.BOOL)
-                .toSignatureString())
-        .isEqualTo("() -> bool");
-    // all kinds of params
-    assertThat(
-            Types.callable(
-                    /* parameterNames= */ ImmutableList.of("x", "a", "b", "c", "d"),
-                    /* parameterTypes= */ ImmutableList.of(
-                        Types.BOOL, Types.INT, Types.FLOAT, Types.NONE, Types.ANY),
-                    /* numPositionalOnlyParameters= */ 1,
-                    /* numPositionalParameters= */ 3,
-                    /* mandatoryParams= */ ImmutableSet.of("a", "c", "x"),
-                    /* varargsType= */ Types.INT,
-                    /* kwargsType= */ Types.INT,
-                    Types.BOOL)
-                .toSignatureString())
-        .isEqualTo(
-            "(bool, /, a: int, b: [float], *args: int, c: None, d: [Any], **kwargs: int) -> bool");
-  }
-
-  private StarlarkType evalType(String type) throws Exception {
-    Expression typeExpr = Expression.parseTypeExpression(ParserInput.fromLines(type));
-    // TODO: #27728 - When type resolution can consider non-universal types, use a better mock
-    // module here that supports evalType().
-    return TypeResolver.evalTypeExpression(typeExpr, Resolver.moduleWithPredeclared());
   }
 }
